requirements.txt
python-telegram-bot
python-telegram-bot[job-queue]
requests
google-generativeai
beautifulsoup4